model_training/requirements.txt
absl-py==1.0.0
anyio==3.5.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
fastapi==0.75.2
Flask==2.0.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h11==0.12.0
h5py==3.6.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.0.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
outcome==1.1.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wrapt==1.14.1
wsproto==1.0.0
